backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pytest==8.3.4
pytest-timeout==2.3.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
